backend/requirements.txt
fastapi
uvicorn[standard]
strawberry-graphql[fastapi]


